线程在强行停止时，先判断是否在运行

+++ src/ui/async_func/async_operate_abc.py
@@ -82,7 +82,8 @@ class ThreadWorkManagerABC(QObject):
         pop_fail(error_msg, self.error_box_title, self.window)
 
     def worker_terminate(self):
-        self.worker.terminate()
+        if self.worker.isRunning():
+            self.worker.terminate()
         self.worker_quit()
         self.post_process()
